Move find_executable to utils
We don't want to import the whole patroni.ctl into the patroni
